Completed protected page and Sign In page
Created the handle sign in and handle sign out function. Also returned it to the page.

Then we created the protected page that displays the users info if user is not null.
If user is null and they try to access the page directly, they will be prompted with a message telling them to log in.